Moved and renamed html parser function
