+++ src/api/one-trust-notices/one-trust-notices.services.ts
@@ -1,17 +1,16 @@
 import { INTEROP_RESOURCES_BASE_URL } from '@/config/env'
-import type { LangCode } from '@/types/common.types'
-import type { RootNode } from '@/utils/parser'
+import type { LangCode, OneTrustContent } from '@/types/common.types'
 import axios from 'axios'
 
 async function getPrivacyPolicyNotice({ lang }: { lang: LangCode }) {
-  const response = await axios.get<RootNode>(
+  const response = await axios.get<OneTrustContent.RootNode>(
     `${INTEROP_RESOURCES_BASE_URL}/one-trust-notices/privacy-policy-${lang}.json`
   )
   return response.data
 }
 
 async function getTermsOfServiceNotice({ lang }: { lang: LangCode }) {
-  const response = await axios.get<RootNode>(
+  const response = await axios.get<OneTrustContent.RootNode>(
     `${INTEROP_RESOURCES_BASE_URL}/one-trust-notices/terms-of-service-${lang}.json`
   )
   return response.data
