+++ src/utils/common.utils.ts
@@ -1,6 +1,6 @@
 import React from 'react'
 import type { TopSideActions } from '@/components/layout/containers/PageContainer'
-import type { ActionItem } from '@/types/common.types'
+import type { ActionItem, OneTrustContent } from '@/types/common.types'
 import type { ButtonProps } from '@mui/material'
 import { getKeys } from '@/utils/array.utils'
 import isEqual from 'lodash/isEqual'
@@ -51,4 +51,57 @@ export function compareObjects<
   return getKeys(object).every((key) =>
     isEqual(object[key], objectToCompare[key as keyof TObjectToCompare])
   )
+}
+
+/**
+ * This function is used to parse the html json to a ReactNode.
+ * We use this approach to avoid the use of dangerouslySetInnerHTML.
+ * @param json the html json
+ * @param route the current route
+ * @returns the ReactNode
+ */
+export const parseHtmlJsonToReactNode = (
+  json: OneTrustContent.Node,
+  route: string
+): React.ReactNode => {
+  function decodeHtml(html: string): string {
+    const txt = document.createElement('textarea')
+    txt.innerHTML = html
+    return txt.value
+  }
+
+  switch (json.node) {
+    case 'root':
+      return React.createElement(
+        React.Fragment,
+        {},
+        json.child.map((item) => parseHtmlJsonToReactNode(item, route))
+      )
+    case 'element':
+      const filteredAttr: Record<string, unknown> = {
+        id: json.attr?.id,
+        href: json.attr?.href,
+        target: json.attr?.target,
+        rel: Array.isArray(json.attr?.rel) ? json.attr?.rel.join(' ') : json.attr?.rel,
+      }
+      /**
+       * During the anchor navigation the root path will be lost.
+       * To avoid that behaviour we put the path behind the anchor in the href
+       */
+      if (filteredAttr.href && (filteredAttr.href as string).startsWith('#', 0)) {
+        if (route.startsWith('/', 0)) {
+          const path = route.substring(1)
+          filteredAttr.href = `${path}${filteredAttr.href}`
+        }
+      }
+      return React.createElement(
+        json.tag,
+        { ...filteredAttr, key: Math.random() },
+        json.child ? json.child.map((item) => parseHtmlJsonToReactNode(item, route)) : undefined
+      )
+    case 'text':
+      return decodeHtml(json.text)
+    default:
+      break
+  }
 }
